+++ src/google/spreadsheets/GameRequestSpreadsheetStyle.ts
